+++ packages/providers-evm/src/ethers.ts
@@ -1,4 +1,4 @@
-import type { ProviderProxy } from '@rarimo/provider'
+import { ProviderProxy, Providers } from '@rarimo/provider'
 import type { providers } from 'ethers'
 
 import { BaseEVMProvider } from '@/base-evm'
@@ -6,5 +6,8 @@ import { BaseEVMProvider } from '@/base-evm'
 export class EthersProvider extends BaseEVMProvider implements ProviderProxy {
   constructor(provider?: providers.Web3Provider) {
     super(provider)
+  }
+  static get providerType(): Providers {
+    return Providers.Ethers
   }
 }